refactor(Instagram): Move the presence extension into the dm package
